try single file

+++ bootstrap.sh
@@ -1,11 +1,13 @@
 sudo apt update
-if ! command -v curl &> /dev/null
-then
-    sudo apt -y remove libcurl4
-    sudo apt -y install curl
-fi
 
-sudo apt -y install git zsh rsync
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    if ! command -v curl &> /dev/null
+    then
+        sudo apt -y remove libcurl4
+        sudo apt -y install curl
+    fi
+    sudo apt -y install git zsh rsync
+fi
 
 ./config-ohmyzsh.sh
